Support multiple chips (#7)
* Adapt to DASv8

* Pipe logs to stdout

This makes them trivially visible in windows PowerShell.

* Add delay to wait for DAS to start

* Generate configuration for memtool instead of relying on existing one

* Implement device scanning and MCD library refactor

Lets start with the ultimate goal of these changes: enable the use of
multiple chips on the same machine with this tool. This required
refactoring of certain parts in the code (we might have killed some bugs
with that as well) and more sophisticated usage of the bindings. Devices
can now be queried and selected specifically, however if only one
device is connected, this is not necessary.

Second error handling is improved. More wrappers are implemented in
rust_mcd::raw; furthermore these wrappers do not return anyhow::Result
anymore but the exact code returned from the API call so that calling
functions can actually check the return type and maybe call the function
again if desired. rust_mcd::raw should remain as simple as possible.

The linux setup for the docker probably compiles, but is completely
untested.

* Fix clippy lints

* Fix formatting

* Inline macro

* Write explicit error message

* Inline macro call

* Remove trailing zero byte

This is added by CString::new.

* Make field private

* Improve error messages

* Make documentation format consistent

* Add information about used field names

* Use correct constructor