feat: Added support for single-core applications and optional elf file (#37)
This PR includes support for not specifying the .elf file on operations that don't require it (eg. --list-devices) and for limiting the number of cores checked for session termination.
In this way it is possible to see defmt messages on single (or dual) core applications without the session terminating abruptly.

The flag --cores has been added to the CLI in order to specify the number of active cores.